Silence two warnings in cmlib. This makes the livecd part compiling warning-free on 32-bit systems. On 64-bit hosts, we still need to squash out this __cdecl warning.
svn path=/trunk/; revision=28851

+++ reactos/lib/cmlib/hiveinit.c
@@ -156,8 +156,8 @@ HvpInitializeMemoryHive(
    if (ChunkSize < sizeof(HBASE_BLOCK) ||
        !HvpVerifyHiveHeader((PHBASE_BLOCK)ChunkBase))
    {
-      DPRINT1("Registry is corrupt: ChunkSize %d < sizeof(HBASE_BLOCK) %d, "
-          "or HvpVerifyHiveHeader() failed\n", ChunkSize, sizeof(HBASE_BLOCK));
+      DPRINT1("Registry is corrupt: ChunkSize %lu < sizeof(HBASE_BLOCK) %lu, "
+          "or HvpVerifyHiveHeader() failed\n", ChunkSize, (SIZE_T)sizeof(HBASE_BLOCK));
       return STATUS_REGISTRY_CORRUPT;
    }
